disturbance: make the logs a common component
